refactor(__init__.py): Pull out code for rendering HTML
